requirements.txt
streamlit
openai
langchain
llama-index
chromadb
pandas
python-dotenv